Add authorization midleware. Fix bugs.

+++ ABAC.WebApp/Controllers/ResourceController.cs
@@ -1,4 +1,5 @@
-﻿using ABAC.DAL.Services.Contracts;
+﻿using ABAC.DAL.Exceptions;
+using ABAC.DAL.Services.Contracts;
 using ABAC.DAL.ViewModels;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -11,10 +12,12 @@ namespace ABAC.WebApp.Controllers
     public class ResourceController : ControllerBase
     {
         private readonly IService<ResourceInfo> service;
+        private readonly IRuleService ruleService;
 
-        public ResourceController(IService<ResourceInfo> service)
+        public ResourceController(IService<ResourceInfo> service, IRuleService ruleService)
         {
             this.service = service;
+            this.ruleService = ruleService;
         }
 
         [HttpGet("")]
@@ -27,6 +30,11 @@ namespace ABAC.WebApp.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetResourceAsync([FromRoute] int id)
         {
+            if (!await ruleService.Validate(HttpContext.Session.GetInt32("userId").GetValueOrDefault(-1), id))
+            {
+                throw new ForbiddenException();
+            }
+
             var result = await service.GetAsync(id);
             return new OkObjectResult(result);
         }
@@ -42,6 +50,11 @@ namespace ABAC.WebApp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteResourceAsync([FromRoute] int id)
         {
+            if (!await ruleService.Validate(HttpContext.Session.GetInt32("userId").GetValueOrDefault(-1), id))
+            {
+                throw new ForbiddenException();
+            }
+
             await service.DeleteAsync(id);
             return new OkResult();
         }
